id-card-detector/requirements.txt
absl-py
astor
cycler
Cython
gast
google-pasta
grpcio
h5py
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
numpy
object-detection-0.1
opencv-python
opt-einsum
Pillow
protobuf
pyparsing
python-dateutil
six
tensorboard
tensorflow
tensorflow-estimator
termcolor
Werkzeug
wrapt
